requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
TikTokApi==6.1.1
playwright==1.41.2
asyncio==3.4.3
httpx==0.24.1
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==21.2.0
nest-asyncio==1.5.8